requirements.txt
fastapi
uvicorn
python-dotenv
openai
gspread
oauth2client
python-multipart
python-docx
pdfminer.six
pandas
numpy